ai/fitting.py
"""
AI Virtual Fitting Service
Uses Replicate API for Stable Diffusion + ControlNet
"""

import os
import replicate
from PIL import Image
import requests
from io import BytesIO
from typing import Optional

class AIFittingService:
    def __init__(self):
        self.api_token = os.getenv("REPLICATE_API_TOKEN")
        if not self.api_token:
            raise ValueError("REPLICATE_API_TOKEN not set")

    async def generate_fitting_image(
        self,
        dog_image_url: str,
        product_image_url: str,
        product_description: str
    ) -> Optional[str]:
        """
        Generate AI fitting image of dog wearing the product

        Args:
            dog_image_url: URL of the dog photo
            product_image_url: URL of the product image
            product_description: Description of the product

        Returns:
            URL of the generated image
        """
        try:
            # Using Stable Diffusion with ControlNet for image generation
            # Model: stability-ai/sdxl with controlnet
            output = replicate.run(
                "stability-ai/sdxl:39ed52f2a78e934b3ba6e2a89f5b1c712de7dfea535525255b1aa35c5565e08b",
                input={
                    "prompt": f"A dog wearing {product_description}, professional product photography, high quality, detailed",
                    "image": dog_image_url,
                    "negative_prompt": "blurry, distorted, low quality, watermark",
                    "num_inference_steps": 30,
                    "guidance_scale": 7.5,
                }
            )

            if output and len(output) > 0:
                return output[0]
            return None

        except Exception as e:
            print(f"Error generating fitting image: {e}")
            return None

    async def generate_fitting_video(
        self,
        dog_image_url: str,
        product_image_url: str,
        product_description: str
    ) -> Optional[str]:
        """
        Generate AI fitting video (future feature)

        Args:
            dog_image_url: URL of the dog photo
            product_image_url: URL of the product image
            product_description: Description of the product

        Returns:
            URL of the generated video
        """
        # TODO: Implement video generation using AnimateDiff or similar
        # For now, return None as placeholder
        return None

    def download_image(self, url: str) -> Image.Image:
        """Download image from URL"""
        response = requests.get(url)
        return Image.open(BytesIO(response.content))


# Alternative: Hugging Face Inference API (Free Tier)
class HuggingFaceAIFitting:
    def __init__(self):
        self.api_token = os.getenv("HUGGINGFACE_API_TOKEN")
        self.api_url = "https://api-inference.huggingface.co/models/stabilityai/stable-diffusion-xl-base-1.0"

    async def generate_fitting_image(
        self,
        dog_image_url: str,
        product_description: str
    ) -> Optional[str]:
        """
        Generate fitting image using Hugging Face Inference API (free tier)
        """
        # TODO: Implement HF inference
        pass
